Merge r374:
BIG fix... The 'msg' field contains the actual message number,
as far as how the server marks it. The 'id' field is simply
an array counter. This means that although normally the
2 are in sync, if a message is removed internally, this is no
longer the case, and the 'id' field will not point to the
correct message. Look at below: In the first case, id 8
maps to msg 8. Now look after we delete a message. In this
case id 8 is the 8th element, or msg 9! To fetch the real
message 9 we need to look at msg (which is 9) and not
id (which would be 8, since it's the 8th message).

LIST
+OK 
1 13580
2 40003
3 40697
4 3744
5 41679
6 41856
7 12043
8 2878
9 41497
10 42444
11 1117
12 1133
.
DELE 6
+OK 
LIST
+OK 
1 13580
2 40003
3 40697
4 3744
5 41679
7 12043
8 2878
9 41497
10 42444
11 1117
12 1133
.